refactor(codegen-plugin): Restructure the code
