+++ packages/codegen-plugin/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["./index.ts"]
+  "include": ["src/"]
 }
